pipeline/requirements.txt
catboost
pandas
apache-airflow
numpy
scikit-learn